+++ flatlaf-core/src/test/java/com/formdev/flatlaf/ui/TestFlatStyleableInfo.java
@@ -469,9 +469,11 @@ public class TestFlatStyleableInfo
 
 			//---- icon ----
 
-			"icon.focusWidth", int.class,
+			"icon.focusWidth", float.class,
 			"icon.focusColor", Color.class,
 			"icon.borderWidth", float.class,
+			"icon.selectedBorderWidth", float.class,
+			"icon.disabledSelectedBorderWidth", float.class,
 			"icon.arc", int.class,
 
 			// enabled
@@ -484,23 +486,30 @@ public class TestFlatStyleableInfo
 			// disabled
 			"icon.disabledBorderColor", Color.class,
 			"icon.disabledBackground", Color.class,
+			"icon.disabledSelectedBorderColor", Color.class,
+			"icon.disabledSelectedBackground", Color.class,
 			"icon.disabledCheckmarkColor", Color.class,
 
 			// focused
 			"icon.focusedBorderColor", Color.class,
 			"icon.focusedBackground", Color.class,
-			"icon.selectedFocusedBorderColor", Color.class,
-			"icon.selectedFocusedBackground", Color.class,
-			"icon.selectedFocusedCheckmarkColor", Color.class,
+			"icon.focusedSelectedBorderColor", Color.class,
+			"icon.focusedSelectedBackground", Color.class,
+			"icon.focusedCheckmarkColor", Color.class,
 
 			// hover
 			"icon.hoverBorderColor", Color.class,
 			"icon.hoverBackground", Color.class,
-			"icon.selectedHoverBackground", Color.class,
+			"icon.hoverSelectedBorderColor", Color.class,
+			"icon.hoverSelectedBackground", Color.class,
+			"icon.hoverCheckmarkColor", Color.class,
 
 			// pressed
+			"icon.pressedBorderColor", Color.class,
 			"icon.pressedBackground", Color.class,
-			"icon.selectedPressedBackground", Color.class
+			"icon.pressedSelectedBorderColor", Color.class,
+			"icon.pressedSelectedBackground", Color.class,
+			"icon.pressedCheckmarkColor", Color.class
 		);
 	}
 
@@ -1040,9 +1049,11 @@ public class TestFlatStyleableInfo
 
 	private void flatCheckBoxIcon( Map<String, Class<?>> expected ) {
 		expectedMap( expected,
-			"focusWidth", int.class,
+			"focusWidth", float.class,
 			"focusColor", Color.class,
 			"borderWidth", float.class,
+			"selectedBorderWidth", float.class,
+			"disabledSelectedBorderWidth", float.class,
 			"arc", int.class,
 
 			// enabled
@@ -1055,23 +1066,30 @@ public class TestFlatStyleableInfo
 			// disabled
 			"disabledBorderColor", Color.class,
 			"disabledBackground", Color.class,
+			"disabledSelectedBorderColor", Color.class,
+			"disabledSelectedBackground", Color.class,
 			"disabledCheckmarkColor", Color.class,
 
 			// focused
 			"focusedBorderColor", Color.class,
 			"focusedBackground", Color.class,
-			"selectedFocusedBorderColor", Color.class,
-			"selectedFocusedBackground", Color.class,
-			"selectedFocusedCheckmarkColor", Color.class,
+			"focusedSelectedBorderColor", Color.class,
+			"focusedSelectedBackground", Color.class,
+			"focusedCheckmarkColor", Color.class,
 
 			// hover
 			"hoverBorderColor", Color.class,
 			"hoverBackground", Color.class,
-			"selectedHoverBackground", Color.class,
+			"hoverSelectedBorderColor", Color.class,
+			"hoverSelectedBackground", Color.class,
+			"hoverCheckmarkColor", Color.class,
 
 			// pressed
+			"pressedBorderColor", Color.class,
 			"pressedBackground", Color.class,
-			"selectedPressedBackground", Color.class
+			"pressedSelectedBorderColor", Color.class,
+			"pressedSelectedBackground", Color.class,
+			"pressedCheckmarkColor", Color.class
 		);
 	}